moved contents from App.js to PostsPage.js
