+++ instagram/src/components/Searchbar/SearchBar.js
@@ -5,7 +5,7 @@ import './SearchBar.css'
 class SearchBar extends React.Component {
     constructor(props) {
         super(props)
-        console.log(props)
+
     }
 
     render() {
